add nav menu file & refactoring

+++ lib/api/model/account.dart
@@ -6,27 +6,26 @@ class AccountModel {
   String email;
   List<String> acls;
   int balance;
-  int pendingCharges;
+  double pendingCharges;
   String lastPaymentDate;
   int lastPaymentAmount;
 
   AccountModel.fromJson(Map<String, dynamic> json)
-    : name = json['name'],
-      email = json['email'],
-      acls = new List<String>.from(json['acls']),
-      balance = json['balance'],
-      pendingCharges = json['pending_charges'],
-      lastPaymentDate = json['last_payment_date'],
-      lastPaymentAmount = json['last_payment_amount'];
+      : name = json['name'],
+        email = json['email'],
+        acls = new List<String>.from(json['acls']),
+        balance = json['balance'],
+        pendingCharges = json['pending_charges'],
+        lastPaymentDate = json['last_payment_date'],
+        lastPaymentAmount = json['last_payment_amount'];
 
   Map<String, dynamic> toJson() => {
-    "name": name,
-    "email": email,
-    "acls": acls,
-    "balance": balance,
-    "pending_charges": pendingCharges,
-    "last_payment_date": lastPaymentDate,
-    "last_payment_amount": lastPaymentAmount
-  };
+        "name": name,
+        "email": email,
+        "acls": acls,
+        "balance": balance,
+        "pending_charges": pendingCharges,
+        "last_payment_date": lastPaymentDate,
+        "last_payment_amount": lastPaymentAmount
+      };
 }
-
